requirements.txt
# Python Face Recognition Microservice - Dependencies
# Install: pip install -r requirements.txt

flask>=3.0.0
face_recognition>=1.3.0
numpy>=1.24.0
Pillow>=10.0.0
dlib>=19.24.0
